Datalogging and Error Checking Implemented (but not tested)
Data is loggged to csv files, openeable by Excel, in a Testbench_log
subdirectory created in the directory where the Testbench code is. Note
that sometimes when the GUI fails when running the DAQ (which happens
alot), the files end up being empty.

Errors are printed to a textbox in the gui, haven't tested any errors besides
FC_ERR_TEMP_H since the temperature values are high when the analog pins
are floating.